requirements.txt
opencv-python
pillow
mediapipe
google-generativeai
streamlit
streamlit-extras
numpy
requests
python-dotenv
